# NSM/auth-service-openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Authentication Service API
  version: "1.0.0"
  description: >
    REST API for the Authentication Service described in auth_service_schema.sql
    and auth-service-database-schema.md. Covers human and machine identity,
    RBAC (roles, permissions, groups), sessions and refresh tokens, API keys,
    service accounts, and read-only audit/login history feeds.


    ## Conventions
    - All request/response bodies are `application/json`, fields are `snake_case`
      (matches the underlying table/column names 1:1 for easy tracing).
    - All timestamps are ISO 8601 / RFC 3339 UTC (`2026-08-07T14:32:00Z`).
    - All resource identifiers are UUIDv4 strings, except `login_history.id` and
      `audit_logs.id`, which are opaque strings backed by a `BIGSERIAL`.
    - The acting tenant is inferred from the caller's access token (`org_id` claim);
      no endpoint accepts an arbitrary `organization_id` override.
    - List endpoints are cursor-paginated: `GET /v1/resource?limit=20&cursor=...`.
    - `POST` endpoints that create a resource accept an optional `Idempotency-Key`
      header; replaying the same key returns the original response instead of
      creating a duplicate.
    - Every endpoint's required permission is a `resource:action` pair, matching
      the `permissions` table's generated `name` column (e.g. `user:write`).
  contact:
    name: Platform Identity Team
servers:
  - url: https://auth.example.com/v1
...
security:
  - bearerAuth: []
paths:
...
  /secrets:
    get:
      tags: [Secrets]
      summary: List secrets (metadata only)
      description: Requires secrets:list. Never returns a value, ciphertext, nonce, or key — see components.schemas.Secret.
      operationId: listSecrets
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Secret' } }
                  page: { $ref: '#/components/schemas/PageMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Secrets]
      summary: Create a secret
      description: Requires secrets:create. The entire `data` object is encrypted as one unit — there is no partial/selective encryption of individual fields.
      operationId: createSecret
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SecretCreate' }
            example: { path: "prod/example-service/db", data: { username: "example_user", password: "REPLACE_ME_EXAMPLE_ONLY", host: "db.internal.example" } }
      responses:
        '201':
          description: Created.
          content: { application/json: { schema: { $ref: '#/components/schemas/Secret' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: A secret already exists at this path.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /secrets/{path}:
    parameters:
      - name: path
        in: path
        required: true
        schema: { type: string }
        description: >-
          The secret's logical, hierarchical path, e.g. `prod/database`.
          May contain `/` as a namespace separator — the entire trailing
          segment is one identifier, never a nested sub-resource route,
          and is never resolved against a filesystem. `.` and `..`
          segments, empty segments, and out-of-range characters are
          rejected (422).
        example: prod/example-service/db
    get:
      tags: [Secrets]
      summary: Read a secret's current or a specific historical version
      description: >-
        Requires secrets:read. This is the only Secrets endpoint whose
        response contains a decrypted value — returned only because this
        specific request already passed server-side authorization.
      operationId: getSecret
      parameters:
        - name: version
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
          description: A specific historical version. Omit for the current version. Versions are immutable and sequential starting at 1.
      responses:
        '200':
          description: OK.
          content: { application/json: { schema: { $ref: '#/components/schemas/SecretValue' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    put:
      tags: [Secrets]
      summary: Create a new version of a secret
      description: >-
        Requires secrets:update. Always creates a new, immutable version —
        never overwrites an existing one. Optimistic concurrency is
        mandatory: If-Match must name the version this update expects to
        be current.
      operationId: updateSecret
      parameters:
        - name: If-Match
          in: header
          required: true
          schema: { type: string }
          description: 'The version this update expects to still be current, e.g. "3". Required on every request — there is no way to update a secret without stating what version it is replacing.'
          example: '"3"'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SecretUpdate' }
            example: { data: { username: "example_user", password: "REPLACE_ME_EXAMPLE_ONLY_2" } }
      responses:
        '200':
          description: A new version was created.
          content: { application/json: { schema: { $ref: '#/components/schemas/Secret' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: If-Match does not equal the secret's current version — re-read the current version and retry; never silently overwritten.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Secrets]
      summary: Soft-delete a secret
      description: Requires secrets:delete. Ciphertext is retained (soft delete only — see components.schemas.Secret); a normal GET afterward returns 404.
      operationId: deleteSecret
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ============================== SERVICE ACCOUNTS ==============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/auth/login`, `/auth/token/refresh`, or `/service-accounts/{id}/token`.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Raw API key, used only to call `/service-accounts/{id}/token` (or directly on resource endpoints the key's scopes allow).
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: UNAUTHENTICATED, message: "Access token is missing or invalid.", request_id: "req_01J9X..." } }
    Forbidden:
      description: Authenticated, but the policy engine denied this action.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: FORBIDDEN, message: "Missing required permission: user:write.", request_id: "req_01J9X..." } }
    NotFound:
      description: Resource does not exist, or is outside the caller's organization.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: NOT_FOUND, message: "No such resource.", request_id: "req_01J9X..." } }
    ValidationError:
      description: Field-level validation failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: VALIDATION_ERROR
              message: "Request failed validation."
              details:
                - { field: "email", issue: "must be a valid email address" }
                - { field: "password", issue: "must be at least 12 characters and include upper, lower, digit, and symbol" }
              request_id: "req_01J9X..."
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    ErrorBody:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          enum:
            - MALFORMED_REQUEST
            - VALIDATION_ERROR
            - UNAUTHENTICATED
            - INVALID_CREDENTIALS
            - MFA_REQUIRED
            - ACCOUNT_LOCKED
            - ACCOUNT_DISABLED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - OWNER_CONFLICT
            - TOKEN_EXPIRED
            - TOKEN_REUSE_DETECTED
            - RATE_LIMITED
            - INTERNAL_ERROR
        message: { type: string }
        details:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/FieldError' }
        request_id: { type: string, example: "req_01J9X8V2QYUR6EZC1G5N3B7K4M" }
    FieldError:
      type: object
      required: [field, issue]
      properties:
        field: { type: string }
        issue: { type: string }

    # ---------- Pagination ----------
    PageMeta:
      type: object
      required: [has_more, limit]
      properties:
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
        limit: { type: integer }
...
    SecretCreate:
      type: object
      required: [path, data]
      properties:
        path:
          type: string
          description: Hierarchical secret path. See /secrets/{path}'s own parameter description for the exact rules.
          example: prod/example-service/db
        data:
          type: object
          additionalProperties: { type: string }
          description: The entire structured secret value — encrypted as one unit, never per-field.
          example: { username: "example_user", password: "REPLACE_ME_EXAMPLE_ONLY", host: "db.internal.example" }
    SecretUpdate:
      type: object
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: { type: string }
          example: { username: "example_user", password: "REPLACE_ME_EXAMPLE_ONLY_2" }
    Secret:
      type: object
      description: >-
        Secret metadata — the response shape for list/create/update. Never
        contains a value, ciphertext, nonce, or encryption key; no field on
        this schema could hold one.
      properties:
        path: { type: string, example: prod/example-service/db }
        version: { type: integer, example: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SecretValue:
      type: object
      description: >-
        A decrypted secret value — returned only by GET /secrets/{path},
        and only after that request already passed server-side
        secrets:read authorization.
      properties:
        path: { type: string, example: prod/example-service/db }
        version: { type: integer, example: 1 }
        data:
          type: object
          additionalProperties: { type: string }
          example: { username: "example_user", password: "REPLACE_ME_EXAMPLE_ONLY", host: "db.internal.example" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ---------- Service Accounts ----------
...